backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
langchain-upstage==0.7.7
pillow==11.3.0
pdf2image==1.17.0
python-dotenv==1.0.1
